Centralize "class like" children related logic
